+++ ios/polar/Sources/polar/PolarPlugin.swift
@@ -586,6 +586,18 @@ public class PolarPlugin:
       data: [identifier, String(describing: feature)])
   }
 
+  public func bleSdkFeaturesReadiness(
+    _ identifier: String, ready: [PolarBleSdkFeature], unavailable: [PolarBleSdkFeature]
+  ) {
+    success(
+      "sdkFeaturesReadiness",
+      data: [
+        identifier,
+        ready.map { String(describing: $0) },
+        unavailable.map { String(describing: $0) },
+      ])
+  }
+
   public func disInformationReceived(_ identifier: String, uuid: CBUUID, value: String) {
     success(
       "disInformationReceived", data: [identifier, uuid.uuidString, value])